requirements.txt
numpy>=1.21
pandas>=1.3
scipy>=1.7
matplotlib>=3.5
plotly>=5.4
panel>=0.13
fastapi>=0.88
uvicorn>=0.20
Pillow>=10.0
